Use WorkRunners while constructing loop
